3: [Susana] Moving all config to single file and BASEDIR out of config object
