+++ book-domain-service/config.py
@@ -1,5 +1,7 @@
 import os
 
+BASEDIR = os.path.abspath(os.path.dirname(__file__))
+
 class Config(object):
 
     SQLALCHEMY_DATABASE_URI = 'sqlite:////tmp/test.db'
@@ -10,3 +12,15 @@ class Config(object):
     def init_app(app):
         pass
 
+
+class Config_Test(object):
+
+    SQLALCHEMY_DATABASE_URI = 'sqlite:////' + os.path.join(BASEDIR, 'test/test_database.db')
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+
+    TESTING = True
+    DEBUG = True
+
+    @staticmethod
+    def init_app(app):
+         pass
